+++ app/src/main/java/com/ecommerce/app/dto/reviews/ReviewsDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class ReviewsDTO {
-    private Long id;
+    private Integer id;
     private int avaliacao;
-    private Long id_produto;
+    private Integer id_produto;
     private String username;
 }
