+++ app/src/main/java/com/ecommerce/app/dto/user/UserDTO.java
@@ -7,11 +7,11 @@ import lombok.Data;
 @Data
 @Builder
 public class UserDTO {
-    private Long id;
+    private Integer id;
     private String username;
     private String email;
 
-    public UserDTO(Long id, String username, String email) {
+    public UserDTO(Integer id, String username, String email) {
         this.id = id;
         this.username = username;
         this.email = email;
