+++ app/src/main/java/com/ecommerce/app/repository/coupons/CouponsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ecommerce.app.model.coupons.Coupons;
 
-public interface CouponsRepository extends JpaRepository<Coupons, Long>{
+public interface CouponsRepository extends JpaRepository<Coupons, Integer>{
 
 	Coupons findByCodigo(String codigo);
